% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropoutSimulation.R
\docType{class}
\name{DropoutSimulation-class}
\alias{DropoutSimulation}
\alias{DropoutSimulation-class}
\title{The DropoutSimulation class}
\description{
This class has slots to store results of different simulations
using Lasso and kknn methods.
}
\section{Slots}{

\describe{
\item{\code{dropout.percentage}}{Vector containing values between 0 and 1
representing the percentages of dropout to simulate over repitions}

\item{\code{n}}{an integer specifying the number of repititions for
a particular dropout percentage}

\item{\code{simulation.result.genes}}{a list storing results of all simulations
where genes are imputed as function of cells}

\item{\code{simulation.result.cells}}{a list storing results of all simulations
where cells are imputed as functions of genes}
}}

