% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropoutSimulation.R
\docType{class}
\name{DropoutSimulation}
\alias{DropoutSimulation}
\alias{plot}
\alias{plot,DropoutSimulation-method}
\alias{simulateDropoutCells}
\alias{simulateDropoutCells,DropoutSimulation-method}
\alias{simulateDropoutGene}
\alias{simulateDropoutGene,DropoutSimulation-method}
\title{The DropoutSimulation class}
\usage{
simulateDropoutGene(theObject, expressionData, dropoutPercentage, n)

\S4method{simulateDropoutGene}{DropoutSimulation}(theObject, expressionData,
  dropoutPercentage, n)

simulateDropoutCells(theObject, expressionData, dropoutPercentage, n)

\S4method{simulateDropoutCells}{DropoutSimulation}(theObject, expressionData,
  dropoutPercentage, n)

plot(theObject, type = "compare", p = 0)

\S4method{plot}{DropoutSimulation}(theObject, type = "compare", p = 0)
}
\arguments{
\item{theObject}{an object of the DropoutSimulation class}

\item{expressionData}{the data matrix containing gene expression values,
genes as rows and cells as columns}

\item{dropoutPercentage}{a vector specifying dropout percentages to simulate}

\item{n}{the number of simulations for each dropout}

\item{type}{the type of graph to plot, "compare" will plot difference of mse(also the default value)}

\item{p}{the dropout percentage to plot, default: 0 will consider the whole range}

\item{theObject}{an object of the DropoutSimulation class}

\item{expressionData}{the data matrix containing gene expression values,
genes as rows and cells as columns}

\item{dropoutPercentage}{a vector specifying dropout percentages to simulate}

\item{n}{the number of simulations for each dropout}

\item{theObject}{an object of the DropoutSimulation class containing simulation results}
}
\description{
This class has slots to store results of different simulations
using Lasso and kknn methods.

Simulations on genes with dependency on cells

Simulations on cells with dependency on genes

This function plots graphs to help analyse between lasso and knn approaches.
}
\section{Slots}{

\describe{
\item{\code{dropout.percentage}}{Vector containing values between 0 and 1
representing the percentages of dropout to simulate over repitions}

\item{\code{n}}{an integer specifying the number of repititions for
a particular dropout percentage}

\item{\code{simulation.result.genes}}{a list storing results of all simulations
where genes are imputed as function of cells}

\item{\code{simulation.result.cells}}{a list storing results of all simulations
where cells are imputed as functions of genes}
}}

