% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn-functions.R
\name{kknnImpute}
\alias{kknnImpute}
\title{kknn imputation - single column}
\usage{
kknnImpute(id, simData.learn, simData.test)
}
\arguments{
\item{id}{string for the cell or gene to impute, must be a column name in the data}

\item{simData.learn}{training data as data.frame}

\item{simData.test}{testing data, for predictions, as data.frame}
}
\value{
data.frame containing results of the simulation
     mean squared error, Spearman correlation, optimal parameters for kknn
}
\description{
Impute results for a specific gene or cell using kknn
}

