% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn-functions.R
\name{kknnImputeAll}
\alias{kknnImputeAll}
\title{kknn imputation - multiple columns}
\usage{
kknnImputeAll(selectedList, simData.learn, simData.test)
}
\arguments{
\item{selectedList}{vector of strings for the cells or genes to impute, must be a column name in the data}

\item{simData.learn}{training data as data.frame}

\item{simData.test}{testing data, for predictions, as data.frame}
}
\value{
data.frame containing results of the simulation
     mean squared error, Spearman correlation, optimal parameters for kknn
}
\description{
Impute results for multiple genes or cells in testing data using kknn
}

