% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedSpeciesLassoNegbin.R
\name{lasso.negbin.mixed.data}
\alias{lasso.negbin.mixed.data}
\title{lasso imputation - single column}
\usage{
lasso.negbin.mixed.data(ID, simData, genes = TRUE)
}
\arguments{
\item{ID}{string for the cell or gene to impute, must be a column name in the data}

\item{simData}{data}

\item{genes}{TRUE/FALSE depending on wether genes are analysed}
}
\value{
data.frame containing results of the simulation
     mean error, percentage of humans
}
\description{
Impute results for a specific gene or cell using lasso
}

