% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso-functions.R
\name{lassoImpute}
\alias{lassoImpute}
\title{lasso imputation - single column}
\usage{
lassoImpute(ID, simData.learn, simData.test)
}
\arguments{
\item{ID}{string for the cell or gene to impute, must be a column name in the data}

\item{simData.learn}{training data, as matrix}

\item{simData.test}{testing data, for predictions, as matrix}
}
\value{
data.frame containing results of the simulation
     mean squared error, Spearman correlation, optimal parameters for kknn
}
\description{
Impute results for a specific gene or cell using lasso
}

