% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso-functions.R
\name{lassoImputeAll}
\alias{lassoImputeAll}
\title{lasso imputation - multiple columns}
\usage{
lassoImputeAll(selectedList, simData.learn, simData.test)
}
\arguments{
\item{selectedList}{vector of strings for the cells or genes to impute, must be a column name in the data}

\item{simData.learn}{training data, as matrix}

\item{simData.test}{testing data, for predictions, as matrix}
}
\value{
data.frame containing results of the simulation
     mean squared error, Spearman correlation, optimal parameters for kknn
}
\description{
Impute results for multiple genes or cells using lasso
}

