% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedSpeciesImput.R
\docType{class}
\name{mixedSpeciesImput-class}
\alias{mixedSpeciesImput}
\alias{mixedSpeciesImput-class}
\title{The mixedSpeciesImput class}
\description{
This class has slots to store results of different simulations
using Lasso, Random Forest and kknn methods.
}
\section{Slots}{

\describe{
\item{\code{simulation.result.genes}}{a list storing results of all simulations
where genes are imputed as function of cells}

\item{\code{simulation.result.cells}}{a list storing results of all simulations
where cells are imputed as functions of genes}
}}

