% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmcClass.R
\docType{class}
\name{PBMCGenes-class}
\alias{PBMCGenes}
\alias{PBMCGenes-class}
\title{The PBMC class}
\description{
This class has methods for the purpose of studying the pbmc dataset for
gene correlation.
}

