% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest_function.R
\name{randomForestImpute}
\alias{randomForestImpute}
\title{kknn imputation - single column}
\usage{
randomForestImpute(id, simData.learn, simData.test, mixedSpeciesData = FALSE)
}
\arguments{
\item{id}{string for the cell or gene to impute, must be a column name in the data}

\item{simData.learn}{training data as data.frame}

\item{simData.test}{testing data, for predictions, as data.frame}
}
\value{
data.frame containing results of the simulation
     mean squared error, Spearman correlation
}
\description{
Impute results for a specific gene or cell using randomForest
}

