% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropoutSimulation.R
\docType{class}
\name{DropoutSimulation}
\alias{DropoutSimulation}
\alias{plot.cells}
\alias{plot.cells,DropoutSimulation-method}
\alias{plot.genes}
\alias{plot.genes,DropoutSimulation-method}
\alias{plot.predict.gene}
\alias{plot.predict.gene,ANY-method}
\alias{simulateDropoutCells}
\alias{simulateDropoutCells,DropoutSimulation-method}
\alias{simulateDropoutGene}
\alias{simulateDropoutGene,DropoutSimulation-method}
\title{The DropoutSimulation class}
\usage{
simulateDropoutGene(theObject, expressionData, dropoutPercentage, n)

\S4method{simulateDropoutGene}{DropoutSimulation}(theObject, expressionData,
  dropoutPercentage, n)

simulateDropoutCells(theObject, expressionData, dropoutPercentage, n)

\S4method{simulateDropoutCells}{DropoutSimulation}(theObject, expressionData,
  dropoutPercentage, n)

plot.predict.gene(expressionData, drop.percent, geneID)

\S4method{plot.predict.gene}{ANY}(expressionData, drop.percent, geneID)

plot.cells(theObject, type = "compare", p = 0)

\S4method{plot.cells}{DropoutSimulation}(theObject, type = "compare", p = 0)

plot.genes(theObject, type = "compare", p = 0, ...)

\S4method{plot.genes}{DropoutSimulation}(theObject, type = "compare", p = 0,
  originalData = NULL, methodAttr = NULL)
}
\arguments{
\item{theObject}{an object of the DropoutSimulation class}

\item{expressionData}{the data matrix containing gene expression values,
genes as rows and cells as columns}

\item{dropoutPercentage}{a vector specifying dropout percentages to simulate}

\item{n}{the number of simulations for each dropout}

\item{geneID}{the geneID to plot}

\item{type}{the type of graph to plot, "compare" will plot difference of mse(also the default value)}

\item{p}{the dropout percentage to plot, default: 0 will consider the whole range}

\item{theObject}{an object of the DropoutSimulation class}

\item{expressionData}{the data matrix containing gene expression values,
genes as rows and cells as columns}

\item{dropoutPercentage}{a vector specifying dropout percentages to simulate}

\item{n}{the number of simulations for each dropout}

\item{expressionData}{the data of gene expression values}

\item{theObject}{an object of the DropoutSimulation class containing simulation results}

\item{theObject}{an object of the DropoutSimulation class containing simulation results}

\item{type}{the type of graph to plot, "compare" will plot difference of mse(also the default value)}

\item{p}{the dropout percentage to plot, default: 0 will consider the whole range}
}
\description{
This class has slots to store results of different simulations
using Lasso and kknn methods.

Simulations on genes with dependency on cells

Simulations on cells with dependency on genes

This function plots the observed v/s predicted values for gene imputation.

This function plots graphs to help analyse between lasso and knn approaches for cell-based imputation.

This function plots graphs to help analyse between lasso and knn approaches for gene-based imputation.
}
\section{Slots}{

\describe{
\item{\code{dropout.percentage}}{Vector containing values between 0 and 1
representing the percentages of dropout to simulate over repitions}

\item{\code{n}}{an integer specifying the number of repititions for
a particular dropout percentage}

\item{\code{simulation.result.genes}}{a list storing results of all simulations
where genes are imputed as function of cells}

\item{\code{simulation.result.cells}}{a list storing results of all simulations
where cells are imputed as functions of genes}
}}

